Move database helper into separate class
Reduce strong coupling between database helper and VPN profiles, to prepare for the addition of other tables.
